Token sale#102 (#104)
* implement API

* progress

* generic invoke

* add whitelist test

* continuing

* update gradle

* validation and balances

* finish formatting

* finishing up

* add dev mode

* add minting

* fix backstack on completion

* finalizing

* finish default urls

+++ app/src/test/java/network/o3/o3wallet/ScriptBuilderTests.kt
@@ -72,7 +72,7 @@ class ScriptBuilderTests {
                 65536 to "0300000100000000")
         val neoScript = ScriptBuilder()
         for (key in testCases.keys) {
-            neoScript.pushInt(key)
+            neoScript.pushInt(key.toLong())
             if (neoScript.getScriptHexString().toLowerCase() != testCases[key]) {
                 throw Exception("Test not passed: ${neoScript.getScriptHexString().toLowerCase()} != $testCases[key]")
             }
